v0.17.0 (#21)
### O language

- Implemented the validation of methods invariant assertions.
- Implemented method post-condition assertions validations.
- Implemented method pre-condition assertions.
- Implemented custom assertions for method calls.
- Rewrote the MethodCallConstraintsValidator to compile a method CompiledAssertion
 on its first use and reuse the same CompiledAssertion from there on.
 Using the tests as a quick benchmark shows an optimization in the code that validates the assertions of a method call of around 150 times.

### FingerTips

### Skins

- TreeView now expands the selection when its set.
- Implemented confirmation dialog.
- Cleaned up references to GtkIcons and moved them to GtkViewsNamespace.
- Implemented a NullViewsNamespace and added protocol to choose which Views library to use.
- Cleaned up NamespaceFlowBuilder, removed duplicated code and moved it to NamespaceFlow.
- Moved the initialization of Gtk to GtkViewsNamespace. Now all the code related with Gtk is encapsulated in GtkViewsNamespace and can be switched to a different implementation by replacing the use of GtkViewsNamespace with a different NamespaceFlow.
- Changed FileLocation to use offsets instead of line-column.

### Sirens

- Implemented a basic browser of spec files.
- Improved how parsers are loaded and filtered.
- Added an icon for JsComment.
- If a TextualContents is a JsComment then uses a JsComment object.
- Implemented JsParserDsl.
- Changed the way file-inspector plugins are loaded to ensure the priority of the plugins. There may be more than one plugin defined for the same file type.
- Remove split-lines as a dependency.
- Cleaned up the classification SourceFile.
- Implemented a YAML files browser.
- Implemented writing a method to its file from the AppBrowser.
- Added confirmation to all documentation deletion actions.
- Changed the initialization of Gtk to not assume any particular GUI library.
- Cleaned up the tests and ensured that no Gtk object is created during the tests of Sirens.
- Splitted the implementation of the FilesRepository classification from its singleton use in the tests. Now tests can use multiple instances without colliding with each other.
- Moved the min-size out from the css files to the ComponentBuilder since it is a functional requirement, not a style.
- Improved and cleaned up the stacks layout definition. Now the layout is done through
 the use of a concrete .spaceFiller() component between consecutive components
 instead of using configuration parameters.

+++ src/sirens/flows/file-object-inspectors/JsClassDocumentationFlow.js
@@ -1,4 +1,5 @@
 const Classification = require('../../../O').Classification
+const Protocol = require('../../../O').Protocol
 const ValueFlow = require('../../../finger-tips/stateful-flows/ValueFlow')
 const EditImplementationNoteDialog = require('../../components/documentation-browser/edition/EditImplementationNoteDialog')
 const EditExampleDialog = require('../../components/documentation-browser/edition/EditExampleDialog')
@@ -11,12 +12,22 @@ const DocumentationImplementationNote = require('../../objects/documentation/sec
 
 const DocumentationIndexSection = require('./DocumentationIndexSection')
 
+
+class JsClassDocumentationFlowMethodAssertions {
+    editClassDocumentationDescription({ parentWindow: parentWindow }) {
+        this.param( parentWindow ) .isNotNull() .isNotUndefined()        
+    }
+}
+JsClassDocumentationFlowMethodAssertions = Protocol.define(JsClassDocumentationFlowMethodAssertions)
+
+
 class JsClassDocumentationFlow {
     /// Definition
 
     static definition() {
         this.instanceVariables = []
         this.assumes = [ValueFlow]
+        this.implements = [JsClassDocumentationFlowMethodAssertions]
     }
 
     /// Building
@@ -278,7 +289,16 @@ class JsClassDocumentationFlow {
         })
     }
 
-    deleteClassDocumentationImplementationNote({ implementationNote: implementationNote }) {
+    deleteClassDocumentationImplementationNote({
+        implementationNote: implementationNote, parentWindow: parentWindow,
+    }) {
+        const confirmsDelete = this.skinsNamespace().ConfirmationDialog.new({
+            confirmationText: 'Delete this implementation note?',
+            window: parentWindow,
+        }).open()
+
+        if( confirmsDelete !== true ) { return }
+
         const classDocumentation = this.getClassDocumentation()
 
         classDocumentation.deleteImplementationNote({
@@ -359,7 +379,14 @@ class JsClassDocumentationFlow {
         })
     }
 
-    deleteClassDocumentationExample({ example: example }) {
+    deleteClassDocumentationExample({ example: example, parentWindow: parentWindow }) {
+        const confirmsDelete = this.skinsNamespace().ConfirmationDialog.new({
+            confirmationText: 'Delete this example?',
+            window: parentWindow,
+        }).open()
+
+        if( confirmsDelete !== true ) { return }
+
         const classDocumentation = this.getClassDocumentation()
 
         classDocumentation.deleteExample({
@@ -381,6 +408,12 @@ class JsClassDocumentationFlow {
     guiNamespace() {
         return this.bubbleUp({
             command: 'guiNamespace',
+        })
+    }
+
+    skinsNamespace() {
+        return this.bubbleUp({
+            command: 'skinsNamespace',
         })
     }
 }
